addDefaultClass HOC
